Add parameter validation
I’ve found that the ESP8266 sometimes maintains old data on the chip
when re-flashed.  Similarly, the REST Post falls over if objectId field
isn’t filled.  When this is complete, it should perform a check to make
sure that all three elements are filled & not run any further if it
doesn’t have any new content to use.